Two subtle but important debugs for checkers
1. This took a long time for me to figure it out:
In Agent multi jump, synchronization of graphics animations in Checkers
and the commands being rapidly sent back to back from Java (and a mix of
threading) was causing a removed user move to be invisible, but stay
there and therefore remain even playable. This took me half a day to
figure it out so I just document it here for nostalgic reasons!

2. Another bug in multi jump of agent is fixed that was allowing agent
to go for a different checker at the second jump.